backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
PyPDF2==3.0.1
sentence-transformers==2.7.0
scikit-learn==1.4.2
numpy==1.26.4
python-dotenv==1.0.1
requests==2.31.0
langchain-groq==0.3.2
